requirements-api.txt
# ─── Core Framework ──────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0

# ─── Database ────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0

# ─── Authentication & Security ───────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
passlib>=1.7.4

# ─── Validation ──────────────────────────────────────────────────────────────
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# ─── Logging & Monitoring ───────────────────────────────────────────────────
structlog>=24.4.0

# ─── HTTP Client ─────────────────────────────────────────────────────────────
httpx>=0.27.0
requests>=2.32.0

# ─── AI / LLM ───────────────────────────────────────────────────────────────
openai>=2.8.0

# ─── Blockchain ──────────────────────────────────────────────────────────────
web3>=7.0.0

# ─── Environment ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# ─── Payments (optional) ────────────────────────────────────────────────────
stripe>=11.0.0
